refactor(atoms): split extraction profile policy
